Final Code Commit (Ready for Review)
Completely Finished Project. Fixed the table that I had created because when I began to input information, I was inputing it in the wrong order. Also, after I finished the table, I created styles for said table and the subtables within it. I worked with heights and widths and even had to brush up on boxshadows. Regardless, it got done.